Notifications, user panel site and routes
View, controller, and route for the user panel have been created. Notifications to pages have been added after successful registration / reset of the user's password, along with a redirection to the login page. After logging in, the user redirects to the user panel page, but after logging in the administrator is redirected to the admin panel.