Soft deletes
Soft deletes have been implemented. From now, the database shows orders that have been deleted by the user along with the date of removal. The user in the preview doesn't see the orders that were deleted because in the controller added a rule that shows only those orders that have NULL in the deleted_at column.